requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
slowapi==0.1.9
yfinance==0.2.40
tavily-python==0.3.3
langchain==0.2.1
langchain-core==0.2.5
langchain-groq==0.1.5
langchain-huggingface==0.0.3
langchain-community==0.2.1
langgraph==0.1.1
qdrant-client==1.9.1
sentence-transformers==3.0.0
requests==2.32.2
beautifulsoup4==4.12.3
numpy==1.26.4
